extract localized test setup into LocalizedTestCase class
