+++ tests/Localized/LocalizedTestCase.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Tests\Localized;
+
+use Tests\TestCase;
+
+abstract class LocalizedTestCase extends TestCase
+{
+    protected $siteFixturePath = __DIR__.'/../Fixtures/site-localized';
+
+    protected function getEnvironmentSetUp($app)
+    {
+        parent::getEnvironmentSetUp($app);
+
+        $app['config']->set('statamic.editions.pro', true);
+        $app['config']->set('statamic.system.multisite', true);
+    }
+}
